rfp-analyzer/requirements.txt
streamlit
langchain
langchain_community
pdfplumber
python-docx
faiss-cpu
sentence-transformers
accelerate
transformers
torch
bitsandbytes
plotly
python-dotenv
tqdm